# services/order/src/app/Presentation/Http/Api/Docs/Schemas/ErrorResponseSchema.php
<?php

declare(strict_types=1);

namespace App\Presentation\Http\Api\Docs\Schemas;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'ErrorResponse',
    required: ['error'],
    properties: [
        new OA\Property(property: 'error', type: 'string', example: 'Order not found: 550e8400-...'),
    ],
)]
#[OA\Schema(
    schema: 'ValidationErrorResponse',
    required: ['message', 'errors'],
    properties: [
        new OA\Property(property: 'message', type: 'string', example: 'The given data was invalid.'),
        new OA\Property(
            property: 'errors',
            type: 'object',
            example: ['buyerId' => ['The buyer id field is required.']],
        ),
    ],
)]
final class ErrorResponseSchema
{
    private function __construct() {}
}