# services/order/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace Payments — Order Service API",
        "description": "REST API for creating and managing marketplace orders. Built with Laravel 11, Ecotone CQRS, and Doctrine ORM.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/v1/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Places a new order in the system. Emits OrderCreated event.",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order by ID",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderView"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/pay": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Start payment for an order",
                "description": "Moves the order into PaymentProcessing state.",
                "operationId": "startOrderPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationStatusResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/paid": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mark order as paid",
                "description": "Moves the order from PaymentProcessing to Paid.",
                "operationId": "markOrderAsPaid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order marked as paid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationStatusResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "description": "Cancels an order. If the order was paid, it is refunded instead.",
                "operationId": "cancelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Out of stock",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order cancelled or refunded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationStatusResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateOrderRequest": {
                "required": [
                    "buyerId",
                    "sellerId",
                    "items"
                ],
                "properties": {
                    "buyerId": {
                        "type": "string",
                        "example": "buyer-1",
                        "maxLength": 64
                    },
                    "sellerId": {
                        "type": "string",
                        "example": "seller-1",
                        "maxLength": 64
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateOrderItem"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
            "CreateOrderItem": {
                "required": [
                    "productId",
                    "quantity",
                    "amount",
                    "currency"
                ],
                "properties": {
                    "productId": {
                        "type": "string",
                        "example": "product-1",
                        "maxLength": 64
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2,
                        "minimum": 1
                    },
                    "amount": {
                        "description": "Amount in smallest currency unit (kopecks)",
                        "type": "integer",
                        "example": 500,
                        "minimum": 1
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB",
                        "maxLength": 3,
                        "minLength": 3
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Order not found: 550e8400-..."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "buyerId": [
                                "The buyer id field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "CreateOrderResponse": {
                "required": [
                    "orderId"
                ],
                "properties": {
                    "orderId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
            "OperationStatusResponse": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "payment_started"
                    }
                },
                "type": "object"
            },
            "OrderView": {
                "required": [
                    "id",
                    "buyerId",
                    "sellerId",
                    "status",
                    "totalAmount",
                    "currency",
                    "items",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "buyerId": {
                        "type": "string",
                        "example": "buyer-1"
                    },
                    "sellerId": {
                        "type": "string",
                        "example": "seller-1"
                    },
                    "status": {
                        "type": "string",
                        "example": "created",
                        "enum": [
                            "created",
                            "payment_processing",
                            "paid",
                            "cancelled",
                            "refunded"
                        ]
                    },
                    "totalAmount": {
                        "description": "Amount in kopecks",
                        "type": "integer",
                        "example": 1000
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemView"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrderItemView": {
                "required": [
                    "productId",
                    "quantity",
                    "amount",
                    "currency"
                ],
                "properties": {
                    "productId": {
                        "type": "string",
                        "example": "product-1"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "amount": {
                        "type": "integer",
                        "example": 500
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter the token only — the \"Bearer \" prefix is added automatically.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Order lifecycle operations"
        }
    ]
}